+++ apps/ptg-configurator/ptgConfiguratorMain.cpp
@@ -130,6 +130,7 @@ const long ptgConfiguratorframe::idMenuQuit = wxNewId();
 const long ptgConfiguratorframe::idMenuAbout = wxNewId();
 const long ptgConfiguratorframe::ID_STATUSBAR1 = wxNewId();
 //*)
+const long ID_TEXTCTRL_SEL_TRAJ = wxNewId();
 
 BEGIN_EVENT_TABLE(ptgConfiguratorframe, wxFrame)
 //(*EventTable(ptgConfiguratorframe)
@@ -252,8 +253,7 @@ ptgConfiguratorframe::ptgConfiguratorframe(wxWindow* parent, wxWindowID id)
 		Panel1, ID_BUTTON5, _("Load Plugin"), wxDefaultPosition, wxDefaultSize,
 		0, wxDefaultValidator, _T("ID_BUTTON5"));
 	FlexGridSizer4->Add(
-		btnLoadPlugin, 1,
-		wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL, 5);
+		btnLoadPlugin, 1, wxALL | wxALIGN_LEFT | wxALIGN_CENTER_VERTICAL, 5);
 	cbHighlightOnePath = new wxCheckBox(
 		Panel1, ID_CHECKBOX3, _("Highlight one trajectory:"), wxDefaultPosition,
 		wxDefaultSize, 0, wxDefaultValidator, _T("ID_CHECKBOX3"));
@@ -275,6 +275,7 @@ ptgConfiguratorframe::ptgConfiguratorframe(wxWindow* parent, wxWindowID id)
 	FlexGridSizer4->Add(
 		edIndexHighlightPath, 1,
 		wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL, 5);
+
 	cbShowOnlySelectedTraj = new wxCheckBox(
 		Panel1, ID_CHECKBOX4, _("Show only selected traj."), wxDefaultPosition,
 		wxDefaultSize, 0, wxDefaultValidator, _T("ID_CHECKBOX4"));
@@ -282,9 +283,21 @@ ptgConfiguratorframe::ptgConfiguratorframe(wxWindow* parent, wxWindowID id)
 	FlexGridSizer4->Add(
 		cbShowOnlySelectedTraj, 1,
 		wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL, 5);
+
 	FlexGridSizer4->Add(
 		-1, -1, 1, wxALL | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL,
 		5);
+
+	edSelectedTrajCmd = new wxTextCtrl(
+		Panel1, ID_TEXTCTRL_SEL_TRAJ, _("Cmd: (none)"), wxDefaultPosition,
+		wxSize(50, -1), wxTE_READONLY, wxDefaultValidator,
+		_T("ID_TEXTCTRL_SEL_TRAJ"));
+	FlexGridSizer4->Add(
+		edSelectedTrajCmd, 1,
+		wxALL | wxEXPAND | wxALIGN_CENTER_HORIZONTAL | wxALIGN_CENTER_VERTICAL,
+		5);
+	FlexGridSizer4->AddGrowableCol(5);
+
 	FlexGridSizer3->Add(
 		FlexGridSizer4, 1, wxALL | wxEXPAND | wxALIGN_LEFT | wxALIGN_TOP, 0);
 	FlexGridSizer8 = new wxFlexGridSizer(1, 0, 0, 0);
@@ -1074,6 +1087,20 @@ void ptgConfiguratorframe::rebuild3Dview()
 				}
 			}
 		}
+
+		// selected PTG trajectory motion cmd:
+		std::string strMotionCmd;
+		try
+		{
+			const auto cmd =
+				ptg->directionToMotionCommand(edIndexHighlightPath->GetValue());
+			strMotionCmd = cmd->asString();
+		}
+		catch (const std::exception& e)
+		{
+			strMotionCmd = e.what();
+		}
+		edSelectedTrajCmd->SetValue(strMotionCmd);
 
 		// 2D angle to robot head plots:
 		std::vector<double> robotHeadAng_x, robotHeadAng_y,