+++ doc/source/doxygen-docs/changelog.md
@@ -1,7 +1,9 @@
 \page changelog Change Log
 
 # Version 2.3.3: UNRELEASED
-- (None)
+- Changes in applications:
+  - ptg-configurator:
+    - Show selected PTG path output motion command.
 
 # Version 2.3.2: Released Jul 14, 2021
 - Changes in applications: